Moving package to handlers Signed-off-by: Varun Jain <varunudr@amazon.com>
